backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pytesseract
PyPDF2
reportlab
pillow==12.0.0
python-dotenv
pdf2image
psycopg2-binary
